refactor(core/serve): extract optimize job state to another file
